Move some headers and libraries to //base
