+++ builder/optimizations.cc
@@ -18,7 +18,6 @@
 #include "runtime/depends_on.h"
 #include "runtime/evaluate.h"
 #include "runtime/expr.h"
-#include "runtime/util.h"
 
 namespace slinky {
