+++ builder/slide_and_fold_storage.cc
@@ -13,7 +13,6 @@
 #include "runtime/depends_on.h"
 #include "runtime/evaluate.h"
 #include "runtime/expr.h"
-#include "runtime/util.h"
 
 namespace slinky {
