+++ builder/test/copy_pipeline.cc
@@ -2,9 +2,9 @@
 
 #include <numeric>
 
+#include "base/test/bazel_util.h"
 #include "builder/pipeline.h"
 #include "builder/replica_pipeline.h"
-#include "builder/test/bazel_util.h"
 #include "builder/test/context.h"
 #include "builder/test/funcs.h"
 #include "builder/test/util.h"
