+++ builder/test/pyramid.cc
@@ -1,9 +1,9 @@
 #include <gtest/gtest.h>
 
+#include "base/test/bazel_util.h"
 #include "builder/pipeline.h"
 #include "builder/replica_pipeline.h"
 #include "builder/substitute.h"
-#include "builder/test/bazel_util.h"
 #include "builder/test/context.h"
 #include "builder/test/funcs.h"
 #include "builder/test/util.h"
