+++ runtime/BUILD
@@ -22,25 +22,8 @@ cc_library(
         "pipeline.h",
         "print.h",
         "stmt.h",
-        "util.h", 
     ],
-    visibility = ["//visibility:public"],
-)
-
-# This thread pool is not directly used by the runtime, it is only used by tests in this package.
-# It could be used in client code, but the expectation is that the client code implements the
-# thread pool interface with some existing parallelism provider rather than using this implementation.
-cc_library(
-    name = "thread_pool",
-    srcs = ["thread_pool.cc"],
-    hdrs = ["thread_pool.h"],
-    visibility = ["//visibility:public"],
-)
-
-cc_library(
-    name = "chrome_trace",
-    srcs = ["chrome_trace.cc"],
-    hdrs = ["chrome_trace.h"],
+    deps = ["//base"],
     visibility = ["//visibility:public"],
 )
