+++ runtime/expr.h
@@ -1,8 +1,10 @@
 #ifndef SLINKY_RUNTIME_EXPR_H
 #define SLINKY_RUNTIME_EXPR_H
 
+#include "base/arithmetic.h"
+#include "base/ref_count.h"
+#include "base/span.h"
 #include "runtime/buffer.h"
-#include "runtime/util.h"
 
 #include <cstdlib>
 #include <initializer_list>
