+++ runtime/test/evaluate_benchmark.cc
@@ -6,7 +6,7 @@
 
 #include "runtime/evaluate.h"
 #include "runtime/expr.h"
-#include "runtime/thread_pool.h"
+#include "base/thread_pool.h"
 
 namespace slinky {
